Correct legacy pagination layer DCHECK in pre-paint.
Floats inside inlines get special treatment in the legacy engine, in
that their Container() (LayoutObject) is the nearest containing block,
rather than the parent. This needs to be reflected everywhere in the
code. We already have special-code for it in e.g.
LayoutObject::PaintingLayer() and PaintPropertyTreeBuilder::
UpdatePaintingLayer(). The DCHECK also needs to join the party.

LayoutObject::FlowThreadContainingBlock() has special behavior for
column spanners inside nested multicol (it will just return nullptr if
it finds a spanner on its way), which is what caused trouble here.

Bug: 1297118
Change-Id: Ie18adf148f0fa31ec1a10eaaf2efc2c074b07e9d