[WPT][Prerender2] Fix restrictions.html timeout
Before this CL, the test could hang if getCurrentPosition returned error.
This CL modifies the test to resolve the promise with the stringified
error to avoid it from hanging.

In addition, this CL also changes Geolocation::PageVisibilityChanged so
that it would care about invisible->visible cases. Before this CL, the
code path assumed visible->invisible case, where it only needed to
shutdown the notifier updates, which didn't require
UpdateGeolocationConnection calls with actual notifier pointers. This CL
updates it to loop for all notifiers registered, so that it can setup
the timers in invisible->visible case.

Bug: 626703
Change-Id: Ibd8ff599c8e2c51555ae3dbd8a55f4252813c405